finances/src/requirements.txt
flask
yahoo-fin
python-binance
pandas